Resolve owner and project through CoprBuildHelper
If no owner and project is given in the vm_image_build
job resolve them trough the CoprBuildHelper defaults
